added missing license headers, moved test to right package
