+++ src/test/java/io/chatpal/solr/ext/logging/JsonLogMessageTest.java
@@ -1,7 +1,23 @@
-package io.chatpal.solr.ext.handler;
+/*
+ * Copyright 2018 Redlink GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package io.chatpal.solr.ext.logging;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import io.chatpal.solr.ext.logging.JsonLogMessage;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
